+++ src/main/mulmo/handler.ts
@@ -26,7 +26,6 @@ import { getProjectPath } from "../project_manager";
 import { loadSettings } from "../settings_manager";
 
 import { createMulmoScript } from "./scripting";
-import { fetchAndSave } from "./fetch_url";
 import { mulmoActionRunner, mulmoGenerateImage, mulmoGenerateAudio, mulmoReferenceImage } from "./handler_generator";
 import {
   mulmoAudioFiles,
@@ -37,6 +36,7 @@ import {
   mulmoReferenceImagesFile,
   mulmoMultiLinguals,
 } from "./handler_contents";
+import { mulmoImageFetchURL, mulmoReferenceImageFetchURL } from "./handler_image_fetch";
 import { graphaiPuppeteerAgent } from "./handler_graphai";
 import { mulmoCallbackGenerator, getContext } from "./handler_common";
 
@@ -163,61 +163,6 @@ const __mulmoImageUpload = async (
 
   return path.join(dirPath, dirKey, filename);
 };
-export const mulmoImageFetchURL = async (projectId: string, index: number, url: string, webContents: WebContents) => {
-  const dirPath = "fetch_image";
-  const dirKey = String(index);
-  return await __mulmoImageFetchURL(projectId, dirPath, dirKey, url, webContents);
-};
-
-export const mulmoReferenceImageFetchURL = async (
-  projectId: string,
-  dirKey: string,
-  url: string,
-  webContents: WebContents,
-) => {
-  const dirPath = "fetch_ref_image";
-  return await __mulmoImageFetchURL(projectId, dirPath, dirKey, url, webContents);
-};
-
-const __mulmoImageFetchURL = async (
-  projectId: string,
-  dirPath: string,
-  dirKey: string,
-  url: string,
-  webContents: WebContents,
-) => {
-  try {
-    const projectPath = getProjectPath(projectId);
-    const dir = path.resolve(projectPath, dirPath, dirKey);
-
-    const res = await fetchAndSave(url, dir);
-
-    if (res.result) {
-      return {
-        result: true,
-        imageType: res.imageType,
-        path: path.join(dirPath, dirKey, res.filename),
-      };
-    }
-    if (res.error) {
-      webContents.send("progress-update", {
-        projectId,
-        type: "error",
-        data: res.error,
-      });
-    }
-  } catch (error) {
-    webContents.send("progress-update", {
-      projectId,
-      type: "error",
-      data: error,
-    });
-  }
-  return {
-    result: false,
-  };
-};
-
 const mulmoUpdateMultiLingual = async (projectId: string, index: number, data: MultiLingualTexts) => {
   const context = await getContext(projectId);
   const { outputMultilingualFilePath } = getOutputMultilingualFilePathAndMkdir(context);
